+++ packages/mfsu/src/index.ts
@@ -1,2 +1,5 @@
 export * from './constants';
+export * from './esbuildHandlers/autoCssModules';
+// for independent use `esbuild-loader`
+export { esbuildLoader } from './loader/esbuild';
 export * from './mfsu';
